이력서CRUD 기능 완료

+++ src/main/java/site/metacoding/miniproject/service/ResumeService.java
@@ -1,5 +1,6 @@
 package site.metacoding.miniproject.service;
 
+import java.util.ArrayList;
 import java.util.List;
 
 import org.springframework.stereotype.Service;
@@ -11,12 +12,17 @@ import site.metacoding.miniproject.domain.application.Application;
 import site.metacoding.miniproject.domain.application.ApplicationDao;
 import site.metacoding.miniproject.domain.resume.Resume;
 import site.metacoding.miniproject.domain.resume.ResumeDao;
+import site.metacoding.miniproject.dto.resume.ResumeReqDto.ApplicationSaveReqDto;
 import site.metacoding.miniproject.dto.resume.ResumeReqDto.ResumeSaveReqDto;
+import site.metacoding.miniproject.dto.resume.ResumeReqDto.ResumeUpdateMainReqDto;
 import site.metacoding.miniproject.dto.resume.ResumeReqDto.ResumeUpdateReqDto;
+import site.metacoding.miniproject.dto.resume.ResumeRespDto.ApplicationSaveRespDto;
 import site.metacoding.miniproject.dto.resume.ResumeRespDto.MatchingResumeRespDto;
 import site.metacoding.miniproject.dto.resume.ResumeRespDto.ResumeAllRespDto;
 import site.metacoding.miniproject.dto.resume.ResumeRespDto.ResumeDetailRespDto;
+import site.metacoding.miniproject.dto.resume.ResumeRespDto.ResumeMyListRespDto;
 import site.metacoding.miniproject.dto.resume.ResumeRespDto.ResumeSaveRespDto;
+import site.metacoding.miniproject.dto.resume.ResumeRespDto.ResumeUpdateMainRespDto;
 import site.metacoding.miniproject.dto.resume.ResumeRespDto.ResumeUpdateRespDto;
 
 @Slf4j
@@ -27,12 +33,29 @@ public class ResumeService {
     private final ResumeDao resumeDao;
     private final ApplicationDao applicationDao;
 
-    public void 지원하기(Application application) {
-        applicationDao.insert(application);
+    @Transactional
+    public ApplicationSaveRespDto 지원하기(ApplicationSaveReqDto applicationSaveReqDto) {
+        log.debug("디버그0 : " + 0);
+        Application applicationPS = applicationSaveReqDto.toEntity();
+        log.debug("디버그1 : " + 1);
+        applicationDao.insert(applicationPS);
+        log.debug("디버그2 : " + 2);
+        applicationPS = applicationDao.findById(applicationPS.getApplicationId());
+        log.debug("디버그3 : " + 3);
+        return new ApplicationSaveRespDto(applicationPS);
     }
 
-    public void 메인이력서등록(Integer resumeId) {
-        resumeDao.updateMain(resumeId);
+    @Transactional
+    public List<ResumeUpdateMainRespDto> 메인이력서등록(ResumeUpdateMainReqDto resumeUpdateMainReqDto) {
+        Resume resumePS = resumeUpdateMainReqDto.toEntity();
+        resumeDao.updateMain(resumePS.getResumeId());
+        List<Resume> resumeList = resumeDao.findByEmployeeId(resumePS.getEmployeeId());
+        List<ResumeUpdateMainRespDto> resumeUpdateMainRespDtoList = new ArrayList<>();
+        for (Resume resume : resumeList) {
+            resumeUpdateMainRespDtoList.add(new ResumeUpdateMainRespDto(resume));
+        }
+
+        return resumeUpdateMainRespDtoList;
     }
 
     public List<ResumeAllRespDto> findResumeAllList() {
@@ -49,33 +72,53 @@ public class ResumeService {
 
     @Transactional
     public ResumeSaveRespDto 이력서작성(ResumeSaveReqDto resumeSaveReqDto) {
-        resumeDao.insert(resumeSaveReqDto);
-        log.debug("디버그 : " + resumeSaveReqDto.getResumeId());
-        Resume resumePS = resumeDao.findById(resumeSaveReqDto.getResumeId());
-        ResumeSaveRespDto resumeSaveRespDto = new ResumeSaveRespDto(resumePS);
-        return resumeSaveRespDto;
+        Resume resumePS = resumeSaveReqDto.toEntity();
+        resumeDao.insert(resumePS);
+        log.debug("디버그 : " + resumePS.getResumeId());
+        resumePS = resumeDao.findById(resumePS.getResumeId());
+        return new ResumeSaveRespDto(resumePS);
     }
 
+    @Transactional
     public ResumeDetailRespDto 이력서상세보기(Integer resumeId) {
         Resume resumePS = resumeDao.findById(resumeId);
-        ResumeDetailRespDto resumeDetailRespDto = new ResumeDetailRespDto(resumePS);
-        return resumeDetailRespDto;
+        if (resumePS != null) {
+            return new ResumeDetailRespDto(resumePS);
+        } else {
+            throw new RuntimeException("해당" + resumeId + "가 없습니다.");
+        }
+
     }
 
     @Transactional
     public ResumeUpdateRespDto 이력서수정(ResumeUpdateReqDto resumeUpdateReqDto) {
-        resumeDao.update(resumeUpdateReqDto);
-        Resume resumePS = resumeDao.findById(resumeUpdateReqDto.getResumeId());
-        ResumeUpdateRespDto resumeUpdateRespDto = new ResumeUpdateRespDto(resumePS);
-        return resumeUpdateRespDto;
+        Resume resume = resumeUpdateReqDto.toEntity();
+        Resume resumePS = resumeDao.findById(resume.getResumeId());
+        if (resumePS != null) {
+            resumeDao.update(resume);
+            return new ResumeUpdateRespDto(resume);
+        } else {
+            throw new RuntimeException("해당" + resume.getResumeId() + "로 수정을 할 수 없습니다.");
+        }
+
     }
 
-    public List<Resume> 내이력서가져오기(Integer employeeId) {
-        List<Resume> resumePS = resumeDao.findByEmployeeId(employeeId);
-        return resumePS;
+    @Transactional
+    public List<ResumeMyListRespDto> 내이력서가져오기(Integer employeeId) {
+        List<Resume> resumeList = resumeDao.findByEmployeeId(employeeId);
+        List<ResumeMyListRespDto> resumeMyListRespDto = new ArrayList<>();
+        for (Resume resume : resumeList) {
+            resumeMyListRespDto.add(new ResumeMyListRespDto(resume));
+        }
+        return resumeMyListRespDto;
     }
 
     public void 이력서삭제(Integer resumeId) {
-        resumeDao.deleteById(resumeId);
+        Resume resumePS = resumeDao.findById(resumeId);
+        if (resumePS != null) {
+            resumeDao.deleteById(resumeId);
+        } else {
+            throw new RuntimeException("해당" + resumeId + "로 수정을 할 수 없습니다.");
+        }
     }
 }
